services/e2e/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
testcontainers==3.7.1
httpx==0.25.2
docker==6.1.3
pika==1.3.2
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
boto3==1.34.0
moto[s3]==4.2.0
python-dotenv==1.0.0
structlog==23.2.0
aio-pika==9.3.1
Pillow==10.1.0 